openconfig/requirements.txt
pyyaml
munch
pandas
argparse